move state transition tracker to its own file
